Moving DevtoolsLog object to client.py
